move function into state
