+++ InvenTree/web/src/src/pages/Login.tsx
@@ -3,12 +3,12 @@ import { Center, Container, Group, Select, Text, Stack } from '@mantine/core';
 import { useAuth } from '../context/AuthContext';
 import { AuthenticationForm } from '../components/AuthenticationForm';
 import { useLocalState } from '../context/LocalState';
-import { fetchServerSession } from '../App';
 import { useToggle } from '@mantine/hooks';
 import { EditButton } from '../components/items/EditButton';
 import { HostOptionsForm } from '../components/HostOptionsForm';
 import { HostList } from '../context/states';
 import { Trans, t } from '@lingui/macro'
+import { useApiState } from '../context/ApiState';
 
 
 export function Login() {
@@ -22,6 +22,7 @@ export function Login() {
       state.lastUsername
     ]
   );
+  const [fetchApiState] = useApiState((state) => [state.fetchApiState]);
   const hostname =
     hostList[hostKey] === undefined
       ? t`No selection`
@@ -47,7 +48,7 @@ export function Login() {
       useLocalState.setState({ lastUsername: username });
       navigate('/');
     });
-    fetchServerSession();
+    fetchApiState();
   }
   function Register(name: string, username: string, password: string) {
     // TODO: Register
